Move definitions out of header files
Header files should only have function declarations. This change converts the header files to function declarations and moves the function definitions to separate cpp files.
